requirements.txt
pandas
numpy
requests
bs4
PyPDF2
pdfminer.six